Refactor Search functions into single purpose component
