+++ src/components/AdvocatesTable/AdvocatesTable.tsx
@@ -1,8 +1,6 @@
 import { useFilteredAdvocates } from "@/hooks/useFilteredAdvocates";
-import { TextInput } from "../TextInput";
-import { Button } from "../Button";
-import { ChangeEvent } from "react";
 import { FormTitle } from "../FormTitle";
+import { AdvocatesTableSearchForm } from "./AdvocatesTableSearchForm";
 
 interface IAdvocatesTableProps {}
 
@@ -10,30 +8,11 @@ export function AdvocatesTable({}: IAdvocatesTableProps) {
   const { advocates, searchTerm, setSearchTerm, resetSearch } =
     useFilteredAdvocates();
 
-  const onChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const searchTerm = e.target.value;
-    setSearchTerm(searchTerm);
-  };
-
-  const onClick = () => {
-    resetSearch();
-  };
-
   return (
     <div className="border border-gray-500 rounded-xl p-4">
       <FormTitle title="Search" />
       <div className="border-b border-gray-300 px-3 py-2 mb-4">
-        <div className="flex flex-direction-row gap-2">
-          <TextInput
-            label={`Searching for: ${searchTerm}`}
-            id="advocates-search-text"
-            type="text"
-            name="search"
-            placeholder="Search by name, phone, city, specialties..."
-            onChange={onChange}
-          />
-          <Button onClick={onClick}>Reset Search</Button>
-        </div>
+        <AdvocatesTableSearchForm searchTerm={searchTerm} setSearchTerm={setSearchTerm} resetSearch={resetSearch} />
       </div>
       <table className="w-full table-auto border-collapse">
         <thead className="text-xs text-gray-700 border-b border-gray-400 uppercase">
@@ -58,7 +37,7 @@ export function AdvocatesTable({}: IAdvocatesTableProps) {
                 <td className="">
                     <div className="max-w-5xl flex flex-direction-row flex-wrap gap-2 p-2">
                   {advocate.specialties.map((s, i) => (
-                    <div key={i} className="text-sm px-3 py-1 bg-gray-100 rounded-xl">{s}</div>
+                    <div key={i} className="text-xs px-3 py-1 bg-gray-100 rounded-xl">{s}</div>
                   ))}
                   </div>
                 </td>
